requirements.txt
django == 2.0.7
gunicorn
django-heroku
pandas
pymongo
matplotlib
bokeh
django-filter
whitenoise
django-model-utils==3.1.2  
gunicorn